Problem: it's clumsy to specify dependencies
I don't like having to expose autoconf semantics in the project model.
(This also applies to package_dependency, which still needs fixing).
The project model should be a clean abstraction. If e.g. Zyre depends
on CZMQ, then it should say just that, no more. It's up to zproject
to do the rest.

Solution: build knowledge of standard projects into zproject, so the
project.xml can say,

    <use project = "czmq" />

And the rest happens automatically.

- known projects are zyre, czmq, zmq
- easy to add more to class_projects.gsl
- resolves further dependencies automagically (zyre->czmq->zmq)