Problem: lack support for zproto state machine animation
The zproto server/client state machines provide an animation option
which shows the state transitions. This is helpful for debugging any
errors in the state machines. This is enabled by passing -v to the
selftest (which passes a verbose argument to server/client classes).

Solution: add 'make animate' target that does a 'make check' with
verbose option.