Problem: asciidoc: FAILED: manpage document title is mandatory
The initial call of zproject_mkman generates incomplete man pages
which then fail with this error in asciidoc. The workaround is to
regenerate all man pages using zproject_mkman. This is not clean
nor obvious.

Solution: add "clean" target that removes .1 and .3 pages and
regenerates .txt and .doc files from source.

Fixes #131