Problem: 'preludefile' feels a bit rough
CZMQ's prelude file is such a 1-off I'm not sure it's sensible to
expose this as a top-level model concept. Isn't it simpler to add
an <header name = ""> entity? We then include this manually in the
project header, which is hand-written, and we allow N headers in a
project. However if we do turn the prelude into a concept, let's
use a shorter, simpler name.

Solution: use 'prelude' instead of 'preludefile'.