Use BUILD_SHARED_LIBS to specify the library type.
This is useful if the project is integrated in another CMake project
using add_subdirectory().  The libary type (STATIC/SHARED) can be
controlled from the master CMakeFile or from the command line
(-DBUILD_SHARED_LIBS).

This commit removes the explicit 'SHARED' type parameter.  If no type is
given explicitly the type is STATIC or SHARED based on whether the
current value of the variable BUILD_SHARED_LIBS is ON. [1]

[1] http://www.cmake.org/cmake/help/v3.0/command/add_library.html

If BUILD_SHARED_LIBS is not specified shared library is build by default.