Check for libtoolize instead of libtool
zproject_autoconf.gsl has a check for the libtool binary as a mean to
check if libtool is available. But distributions like Debian and
Ubuntu are splitting the libtool package, and the libtool binary is now
in a separate package. What autoconf actually need is not the libtool
binary, but libtoolize and other macro files. So check for libtoolize
instead.