qt-android: set PKG_CONFIG_LIBDIR path to NDK dir
pkg-config will use the default path unless PKG_CONFIG_LIBDIR is set,
which means when doing cross-builds with the Android NDK the host
system pkg-config default path will be used, which is incorrect,
since we must build against the libraries in the DevKit.
Point instead pkg-config to the only NDK path that currently ships
pkg-config files. This might need to be updated if the NDK changes.