Problem: generated bindings failing on Python 3.
Bindings generated for projects that use the buffered I/O (FILE*) C
library calls, e.g. "czmq", will not run on the version 3 of the
interpreter. The actual problem is that the API used to cast files from
Python to C, and vice-versa, is not available on this version. See

  https://docs.python.org/3/c-api/file.html

Solution: update the GSL script to use the Standard C library functions
on Python 3.