Problem: Includes didn't work properly
I noticed that my commit (6d01f82d90 make resolve_includes() reusable)
caused the license text to disappear in some generated files. The
function worked for partly, but not fully.

Solution: Fix it.

One must not use real tabs in GSL scripts or it leads to very strange
errors. If I realized this before, I'd have gone for this solution in
the first place.