Problem: hacky check for foreign types failed in zarmour
The code tested for foreign type dependencies by comparing the C
type with abstract type; zhash -> zhash_t *. If so, it looked for
an external class to resolve the type.

This all works until you pass arguments by reference and use a
type that is implemented as _t in C. The specific case is zarmour
which passes a 'size' argument by reference.

Solution: use an explicit 'is a foreign type' attribute on
containers:

    my.container.foreign = 1

and test for this as needed. Lesson: please don't use hacky lossy
tests that break on unexpected data...