Problem: QT binding corrupts size attribute on numbers
It erroneously uses the container size to store the result of
a regexp, causing a number/4 to become a number/32. The result
is that further attempts to use the container types will fail.

(The Ruby binding avoids this trap by using the C types, not
the container types.)

Solution: use my.size instead of size.