Problem: private attribute is fragile
E.g. if you set 'private = "0"' on an element, it breaks a lot
of code that does stuff like 'if defined (whatever.private)'.

Solution: use explicit scope attribute with values "public" or
"private" so that code is unambiguous and unfragile.

Done for classes, constants, headers, mains, and actors.

Problem #2: selftest main is a special case...

This makes things more complex than they should be, and hints to
target developers that they can ignore other mains. (Leading to
unbuilt mains in various of the targets.)

Solution: always define a private main specifying the selftest.