Problem: CMake builds executables in root directory
This is inconsistent with the automake scripts, and the generated
.gitignore file. It is impossible to ignore executables in the root
directory without ignoring any directory with the same name, which
breaks MSVC builds.

Solution: fix CMake to build binaries in src.

Comment: it might be better to build binaries in bin, which raises
a number of other question such as where do we build libraries, and
do we split public/private binaries? For now, keeping it all in src
is simplest.