Problem: CMake builds files in root directory
This is untidy, and inconsistent with automake scripts. I don't know
if this was deliberate or not. The intention was perhaps to let the
user override the target binary directory, yet do we need this? The
code wasn't even consistent (platform.h is not a binary yet was being
sent to BINARY_DIR as well.

Solution: quash use of BINARY_DIR and generate ALL outputs in
${SOURCE_DIR}/src, in same manner as autoconf scripts do.