Problem: ci_build.sh is artificially frozen
This isn't a file that users should be writing and changing. Yet
it only gets generated once, and is then frozen. The assumption
is that it'll be heavily customized. This appears false. What we
see instead is that it's out of date in projects that have it.

Solution: regenerate every time. If someone needs a hand-written
override, they can modify .travis.yml to call a different file.