Problem: C element status needs to be initialised
Solution: if the API XML exists and defines a status for the C
element, use it to initialise the status variable.
If a child element of a class does not define a status in the API XML
take the parent class value.
If a class does not define a status in the API XML, default to
"draft".