Problem: newly generated class source is no good
The template tries to be clever and use the API model as basis
for the source; however this is overkill and does not work in
the simpler case where we just specify the class in the project
model.

Solution: generate new/destroy/print/test methods to match the
header file.