Problem: can't build CZMQ stable
The library uses its own draft methods internally. The two options seem
to be to check CZMQ_BUILD_DRAFT_API in various places, or to enable the
draft API for internal library use. I tried the first and it is nasty.

Solution: provide prototypes for all draft methods in the internal
project header (src/czmq_classes.h).

Also, improved the @interface documentation for draft and legacy
methods so the man pages are absolutely clear.