Problem: version extraction is much too complex
The version extraction code in the CMake and Autotools scripts
works as we used to by hand, via a version.sh script that reads
the project header.

Since we regenerate the project header and we know the version
explicitly (it's in the project.xml) there is no need for this
balancing act.

Solution: rip it out and stick proper version into generated
scripts.