Problem: zproject does not generate CFFI-based python bindings.
I want these for malamute, and also apparently pyczmq is a bit
under-maintained due to the pain of doing so manually.

Solution: add zproject_python_cffi.gsl, which builds target
python_cffi.  This can be the base for malamute bindings.  It doesn't
fully replace pyczmq by any means, but it's a step in that direction.
Note that this does not yet distinguish between stable/legacy and
draft components.