Problem: .xml extension for API files is ambiguous
Especially if we have other files for a class, e.g. in zproto we
put the state machine models in src/myclass.xml.

Solution: use .api as expenstion for API files

The code will still work with existing .xml files, however the docs
now refer to .api as the correct extension.