Problem: drafts are confusing on many projects
It's clumsy that projects that don't even care about this aspect
are forced to use it, or they don't build properly. That is, the
project header is useless on projects consisting only of draft
classes, which is the default for new projects.

Solution: always build drafts on projects that have no stable
public classes.

Fixes #550