Problem: debian generates dsc file, breaks build
Solution: do not autogenerate a .dsc file in zproject. This not only
is not needed, as the .dsc file is generated when building the
package itself, but it also breaks building in a real-world use case
on the Open Build Service, which issues a recursive dpkg-scansources
that will find this .dsc among the source files, and the build will
abort as it's not expected.