Problem: enums are over-engineered
These are just constants, yet the syntax for defining and using
them is complex and does not solve any identifiable problem except
"maintain compatibility with existing header files".

As well as being complex to specify and use, they're a pain to
implement in bindings.

The simplest solution is to use constants. I've done that in CZMQ
and Zyre and now here in zproject.

Solution: use constants.

Fixes #460