Problem: nodejs binding does not deal with CZMQ
Mainly, needed to add support for more complex arguments and
return types. This now builds a binding for CZMQ, and the result
compiles. It is not complete. I've no idea yet how to return class
types from a method.

Solution: add support for CZMQ types as far as possible here.