Problem: [travis] make memcheck causes failing builds all the time Solution: Remove memcheck from the travis build.
During the last month `make memcheck` caused all zproject based builds to
fail because there are supposedly memory leaks within other layers of the
software e.g. libzmq, which causes the build to fail all the time. IMO
the memcheck should be run manually in addition to testing and not by
travis itself.