Problem: [travis] deploying release artifacts is a manual process
Solution: Use travis to deploy these artifacts to github automatically.

The deployment is triggered by tagging on master branch of the
repository. Travis will trigger deployment on the default build
and will in process disable drafts.

For now the results of `make distcheck` are deployed
as well as their md5 and sha1 hash sums.