Problem: [jni] android cmake does not use pkgconfig to find deps
Solution: Use the same mechanism which is used by the jni bindings and
the project itself. This change also modifies the toolchain file and the
build script accordingly. One improvement of this is that the pkgconfig
won't search for libraries installed on the system but only for the ones
that are compiled for android.