Problem: [jni] android unzip during build fails
Solution: Explicitly state the file to unzip

Previously the regex used in the unzip command did work as it match one
file. Recent changes in the gradle build introduced sources and
javadoc jar files which are located in the same directory. Thus the
unzip command needs to explicitly state the jar file it wants to
extract.