Problem: build failures due to Wredundant-decls
Solution: roll back duplicate declarations for functions that have
a format index parameter.
This will make the docs a bit uglier, by appending the CHECK_PRINTF
macro at the end of the signature in the doc.
But a duplicate declaration in a public header breaks the build of
any project that uses the library and includes them, and also builds
with -Wredundant-decls.

  /usr/local/include/zcert.h:128:5: error: redundant redeclaration
   of ‘zcert_set_meta’ [-Werror=redundant-decls]
     zcert_set_meta (zcert_t *self, const char *name, const char
                     *format, ...) CHECK_PRINTF (3);
     ^
  /usr/local/include/zcert.h:64:5: note: previous declaration of
   ‘zcert_set_meta’ was here
     zcert_set_meta (zcert_t *self, const char *name, const char
                     *format, ...);