Problem: make distcheck is no longer ran by default
Solution: make the new distcheck project config option opt-out
instead of opt-in.
Correct handling and creation of distributable tarball is core to the
github and travis based releases scripts that zproject generates, so
by default it should check that everything works correctly.
Retain the option so that users, if they intend to do so at their
risk, can disable this check in their projects.