Problem: need to automate self-testing of private classes
Solution: add a new module, $project_private_selftest, which will
call every private class selftest function. This module exposes a
single void $project_private_selftest (bool verbose) function which
is then called from the selftest binary.
This way there is a single, common and stable public symbol and
private classes symbols are not exposed.
Note that for now it's exposed as a DRAFT functionality to allow for
testing.
Fixes #808