Problem: can't install additional files via project.xml
Solution: introduce zinstall, the zero model for file installation

The beauty of this model (as opposite of PR#539) is that it impose clear
protocol between model and rest of the generators. This protocol are
files in builds/zinstall/, which are then included in generated files.

This has two main advantages over older attempts
1. installation model fully implemented in ONE place
2. because of the protocol, there can be more models w/o any impact on
backend

Design goals are stated in README.txt, but let's repease
1. KISS - the less options the better
2. No conditionals - please don't create an another half-baked XML
programming langugage like ant
3. Real problems - don't add more options because you can
... or you will get replaced by zinstall2 ;-)