Problem: Generation of systemd service is missing
Solution: Add some complementary information

When using both autotools and systemd, since configure.ac is generated first
and checks for the actual presence of service file to list these in
AC_CONFIG_FILES directive, and since the generation of these service files is
done after the configure.ac generation, it results that systemd service files
generation are missing from configure.ac, and thus the files are not generated
from the template. A workaround is to call twice `gsl project.xml`, but it is
worth a fix.

Signed-off-by: Arnaud Quette <ArnaudQuette@Eaton.com>