Problem: Python CFFI only works with DRAFTS on CZMQ
Solution: parse DEB_BUILD_OPTIONS and the OBS config for the
python_cffi macro and build the bindings only if it's set.
This is similar to what RPM does, but it still creates the emtpy
packages.

Although it is possible to generate dynamically new binary packages
at build time for Debian, it's very hacky and dirty so it's better to
avoid it unless absolutely necessary.