Problem: no support for nopython Debian build profiles
Solution: implement support for "nopython". Build profiles can be
used to select a subset of binary packages to build (and build
dependencies to install), useful for example when deboostrapping a
new architecture or when acting on a more restricted environment
where not all packages are available, for example an embedded
system.
Note that the implementation is partial, as Debian 7 and Ubuntu
12.04 do not support <!nopython> qualifiers in the build dependencies
list, so it cannot be added yet.