Problem: you cannot mix two ffi instances in one call
Solution: let the upper module contains the definitions of all dependant ones.

TL;DR;
The original error code was

    TypeError: initializer for ctype 'struct _mlm_client_t *' appears indeed to
    be 'struct _mlm_client_t *', but the types are different (check that you are
    not e.g. mixing up d ifferent ffi instances)

The point is that every $(project) defines own ffi instance, which sadly can't
be mixed in one function call.

    client = mlm_cffi.lib.mlm_client_new ()
    msg = zm_proto_cffi.lib.zm_proto_new ()
    zm_proto_cffi.lib.zm_proto_recv (msg, client)

zm_proto_recv works with two declarations of mlm_client_t and thus fails. The
problem is that every cffi module must be self-contained, so all types must be
resolved. Proposed solution is kind of static linking, but it have two advantages

1. following code works

    client = zm_proto.lib.mlm_client_new ()
    msg = zm_proto_cffi.lib.zm_proto_new ()
    zm_proto_cffi.lib.zm_proto_recv (msg, client)

2. there is no need to add python dependencies to python2-zm-proto-cffi package