Problem: test listing is not 100 % acurate
Solution: as there are 4 states class can be in (public/private,
stable/draft), there are cases listing is not correct. Solve this via
iterating through array of test_item_t, which now contain flags for
public/private, stable/draft.

This avoids copying nested gsl ifs in the template and --list command is
not performance critical anyway.