fix configuration names in build.bat
Building on windows by calling build.bat fails as it tries to build
projetc configurations which do not exists in the visual studio solution
or project files. This commit changes the configuration names in the genereated
build.bat file to be in sync with the visual studio solution and project
file.