revert d9ae126 (PR 1060)
the project build configuration were not matching the solution file
only after opening the solution file in visual studio. It apperently
did not like the solution file and replaced solution build
configurations and platforms.
So it is better to revert to the original configuration names in the
generated build.bat file and improve the generated solution file.