Problem: Private class self tests hidden by public test failures
Solution: Shuffle the test order, so private classes are tested first.
Assumption is that they are the internal implementation detail,
and if they fail their selftests in a new iteration, the public
classes would also fail for these reason (but without delivering
the real cause of failure so clearly, being higher-level consumers).

Signed-off-by: Jim Klimov <EvgenyKlimov@eaton.com>
Asked-for-by: Jiri Kukacka <JiriKukacka@eaton.com>