Problem: test for C++11 (so g++4.9) is unavoidable
Solution: make the check for std::regex (assuming it means sufficient
level of C++11 in GCC) optional with --enable-gcc-std-regex=no and by
default enabled.

A build with this option set to 'no' should compile, but the result
would not necessarily function (gcc-4.8 and earlier could include the
API routines for supporting the standard, but not really implement it).

Primary intended use-case:

If Travis CI tests of some projects "use" a project that by itself
requires C++11, but do not really use this aspect of it in their own
unit-tests, this is sufficient for "trusty" machines with older GCC.
In this case, the "<use>" tag for such project can be amended with

    <add_config_opts>--enable-gcc-std-regex=no</add_config_opts>

tags to avoid configure-time failures in Travis CI builds of the
whole dependency chain (unpackaged mode).

Signed-off-by: Jim Klimov <EvgenyKlimov@eaton.com>