Problem: default starting_year of 2015 in projects
Solution: if a project is being just generated, and there
is already an <include filename="license.xml"> in project.xml
but no license.xml (verbatim) yet, create one with the content
of <starting_year> (set to today's year) and a <license> tag
with a reminder to define an actual license and regenerate.

Do not touch existing project license files though.

Signed-off-by: Jim Klimov <EvgenyKlimov@eaton.com>