Problem: too easy to release a project with obsolete Copyrights
Solution: Take legalese seriously. During regeneration, check
if the current year is mentioned (from, until, listed, whatever)
in a line with the "Copyright" word or some of the ways it can
be written in the XML `<license>` tag.

If the year is not listed in any such line (or no such lines are
there), warn the user in at least the build-logged messages, and
also pause for 5 sec so the interactive people can break, fix and
regenerate.

Signed-off-by: Jim Klimov <EvgenyKlimov@eaton.com>