Problem: All resolved implied dependencies are taken as project dependencies
While resolving dependencies recursively is a nice feature there are
dependencies that are private for an dependency and the project itself
does not care for. But those private dependencies are important to know
in case we need to build them from source for a ci build.

Solution: Allow to declare dependencies as private

Implied and private dependencies are ignored as dependency but
considered for ci builds.

Includes for private dependencies are now placed in the private classes
header.